+++ lib/routing/router.dart
@@ -1,3 +1,4 @@
+import 'package:ephor/ui/dashboard/supervisor/view_model/dashboard_viewmodel.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
@@ -26,7 +27,7 @@ GoRouter router(AuthRepository authRepository) => GoRouter(
       path: Routes.home,
       builder: (context, state) {
         return DashboardView(
-          viewModel: ""
+          viewModel: DashboardViewModel(authRepository: context.read())
         );
       }
     )
